refactor: extract scoring functions from evals into impact_mappings
Move score_impact_diff, score_cvss3_diff, and their supporting
constants out of evals/features/cve/test_suggest_impact.py into the
shared impact_mappings module so production code can reuse them
without importing from the eval suite.
